Added option to find and replace all occurrences of existing text.
Added a new text box that accepts additional text. If the newly added find and replace checkbox is checked, and the newly added find and replace button is clicked, the program will replace all occurrences of the first text with the second.
Tested (manually) successfully on a deck of about 6k cards.
Lots of scope for improvement.